Fixed /usr/bin/evn typo in all files. Converted to LMCSS notation. Implemeted pymol alignment
